SIS3305: Corrected handling of the sample lengths. Rewrote readout to handle a single waveform per packet and changed variables to reflect this. Changed indentation of the decoder, which caused a huge number of changes to appear on accident.
git-svn-id: svn://orca.physics.unc.edu/Orca@10097 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/Vme/SIS3305/ORSIS3305Model.h
@@ -301,7 +301,7 @@
 
 
 
-- (unsigned short) sampleLength:(short)group;
+- (unsigned long) sampleLength:(short)group;
 - (void) setSampleLength:(short)group withValue:(unsigned long)aValue;
 
 - (int)  triggerGateLength:(short)group;
